Initial upload of ArcXos custom build profiles, configurations, and assets with documentation

+++ slim-iso/airootfs/root/customize_airootfs.sh
@@ -24,7 +24,7 @@ sed -i 's/#\(HandleHibernateKey=\)hibernate/\1ignore/' /etc/systemd/logind.conf
 sed -i 's/#\(HandleLidSwitch=\)suspend/\1ignore/' /etc/systemd/logind.conf
 
 # enable useful services and display manager
-enabled_services=('choose-mirror.service' 'lightdm.service' 'dbus' 'pacman-init'
+enabled_services=('choose-mirror.service' 'lightdm.service' 'pacman-init'
   'NetworkManager' 'irqbalance' 'vboxservice')
 systemctl enable ${enabled_services[@]}
 systemctl set-default graphical.target
@@ -40,14 +40,14 @@ echo "blacklist pcspkr" > /etc/modprobe.d/nobeep.conf
 
 # disable network stuff
 rm -f /etc/udev/rules.d/81-dhcpcd.rules
-systemctl disable dhcpcd sshd rpcbind.service
+systemctl disable dhcpcd sshd rpcbind.service systemd-networkd.service systemd-networkd-wait-online.service systemd-resolved.service iwd.service
 
 # remove special (not needed) files
 rm -f /etc/systemd/system/getty@tty1.service.d/autologin.conf
 rm -f /root/{.automated_script.sh,.zlogin}
 
 # setting root password
-echo "root:blackarch" | chpasswd
+echo "root:arcx" | chpasswd
 
 # copy config files to skel
 cp /usr/share/blackarch/config/bash/bashrc /etc/skel/.bashrc
@@ -56,13 +56,15 @@ cp /usr/share/blackarch/config/zsh/zshrc /etc/skel/.zshrc
 
 # setup user
 useradd -m -g users -G wheel,power,audio,video,storage -s /bin/zsh liveuser
-echo "liveuser:blackarch" | chpasswd
-ln -sf /usr/share/icons/blackarch-icons/apps/scalable/distributor-logo-blackarch.svg /home/liveuser/.face
+echo "liveuser:arcx" | chpasswd
+ln -sf /usr/share/backgrounds/arcxoslogo.png /home/liveuser/.face
 mkdir -p /home/liveuser/Desktop
 chown -R liveuser:users /home/liveuser/Desktop
 chmod -R 755 /home/liveuser/Desktop
-ln -sf /usr/share/applications/calamares.desktop /home/liveuser/Desktop/calamares.desktop
-sed -i -e "s|Install System|Install BlackArch|g" /usr/share/applications/calamares.desktop
+if [ -f /usr/share/applications/calamares.desktop ]; then
+    ln -sf /usr/share/applications/calamares.desktop /home/liveuser/Desktop/calamares.desktop
+    sed -i -e "s|Install System|Install ArcXos|g" /usr/share/applications/calamares.desktop
+fi
 ln -sf /usr/share/applications/xfce4-terminal-emulator.desktop /home/liveuser/Desktop/terminal.desktop
 chmod +x /home/liveuser/Desktop/*.desktop
 
@@ -71,6 +73,9 @@ cp -r /etc/skel/. /root/.
 
 # repo + database
 curl -s https://blackarch.org/strap.sh | sh
+# Ensure we have working mirrors enabled in blackarch-mirrorlist
+sed -i 's|#Server = https://ftp.halifax.rwth-aachen.de|Server = https://ftp.halifax.rwth-aachen.de|g' /etc/pacman.d/blackarch-mirrorlist
+sed -i 's|#Server = https://www.blackarch.org|Server = https://www.blackarch.org|g' /etc/pacman.d/blackarch-mirrorlist
 pacman -Syy --noconfirm
 pacman-key --init
 pacman-key --populate blackarch archlinux
@@ -86,11 +91,15 @@ rm -f /etc/fonts/conf.d/09-autohint-if-no-hinting.conf
 
 # temporary fixes for ruby based tools
 cd /usr/share/whatweb && rm -f Gemfile.lock &&
-  bundle config build.nokogiri --use-system-libraries &&
-  bundle install --path vendor/bundle && rm -f Gemfile.lock
+  bundle config set build.nokogiri --use-system-libraries &&
+  bundle config set path 'vendor/bundle' &&
+  bundle install && rm -f Gemfile.lock
 
 # change default jdk
-archlinux-java set java-20-openjdk
+default_java=$(archlinux-java status | grep -o 'java-[a-zA-Z0-9/-]\+' | head -n 1)
+if [ -n "$default_java" ]; then
+  archlinux-java set "$default_java"
+fi
 
 # Temporary fix for calamares
 #pacman -U --noconfirm https://archive.archlinux.org/packages/d/dosfstools/dosfstools-4.1-3-x86_64.pkg.tar.xz
@@ -99,9 +108,35 @@ archlinux-java set java-20-openjdk
 gdk-pixbuf-query-loaders --update-cache
 
 # /etc
-echo 'BlackArch Linux' > /etc/arch-release
+echo 'ArcXos' > /etc/arch-release
 
 # vim
 cp -r /usr/share/blackarch/config/vim/vim /home/liveuser/.vim
 cp /usr/share/blackarch/config/vim/vimrc /home/liveuser/.vimrc
+
+# Ensure XFCE desktop config sets wallpaper path correctly
+mkdir -p /etc/skel/.config/xfce4/xfconf/xfce-perchannel-xml/
+find /etc/skel/.config/xfce4/ /root/.config/xfce4/ -name "xfce4-desktop.xml" 2>/dev/null | while read -r config_file; do
+  sed -i 's|/usr/share/backgrounds/[^"]*|/usr/share/backgrounds/background.png|g' "$config_file"
+  sed -i 's|/usr/share/blackarch/wallpaper/[^"]*|/usr/share/backgrounds/background.png|g' "$config_file"
+done
+
+# Override any wallpaper in blackarch wallpaper folder if it gets installed
+if [ -d /usr/share/blackarch/wallpaper ]; then
+  find /usr/share/blackarch/wallpaper/ -type f -name "*.png" -o -name "*.jpg" | while read -r wp; do
+    cp /usr/share/backgrounds/background.png "$wp"
+  done
+fi
+
+# Override default XFCE wallpapers to ensure they display our custom background
+if [ -d /usr/share/backgrounds/xfce ]; then
+  find /usr/share/backgrounds/xfce/ -type f | while read -r wp; do
+    cp /usr/share/backgrounds/background.png "$wp"
+  done
+fi
+if [ -d /usr/share/xfce4/backdrops ]; then
+  find /usr/share/xfce4/backdrops/ -type f | while read -r wp; do
+    cp /usr/share/backgrounds/background.png "$wp"
+  done
+fi
